TL-Stream/streaming/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.6.0
python-dotenv==1.0.1
httpx==0.27.0
